Reset update request after sending Improved rendering view of adding asset forms
