+++ views/server-config/add-classic-asset-entry-form.js
@@ -1,23 +1,13 @@
 import React, {useCallback, useRef, useState} from 'react'
-import {Button} from '@stellar-expert/ui-framework'
 import {observer} from 'mobx-react'
 import {StrKey} from 'stellar-sdk'
+import AddAssetEntryLayout from './add-asset-entry-layout'
 
 export default observer(function AddClassicAssetEntry({title, settings, save}) {
-    const [isVisible, setIsVisible] = useState(false)
-    const [isValid, setIsValid] = useState(false)
     const [asset, setAsset] = useState({})
+    const [isValid, setIsValid] = useState(false)
+    const [isEntered, setIsEntered] = useState(false)
     const currentInput = useRef(null)
-
-    const toggleShowForm = useCallback(() => {
-        setIsVisible(!isVisible)
-        setTimeout(() => {
-            const input = currentInput.current
-            if (input) {
-                input.focus()
-            }
-        }, 200)
-    }, [isVisible])
 
     const validate = useCallback(newAsset => {
         const pattern = new RegExp("[^a-zA-z]")
@@ -43,31 +33,12 @@ export default observer(function AddClassicAssetEntry({title, settings, save}) {
         setAsset(newAsset)
         setIsValid(validate(newAsset))
     }, [asset, validate])
-
-    const onSave = useCallback(() => {
-        save(`${asset.code}:${asset.issuer}`)
-        toggleShowForm()
-    }, [asset, save, toggleShowForm])
     //save on "Enter"
-    const onKeyDown = useCallback(function (e) {
-        if (e.keyCode === 13 && isValid) {
-            onSave()
-        }
-    }, [isValid, onSave])
+    const onKeyDown = useCallback(e => setIsEntered(e.keyCode === 13 && isValid), [isValid])
 
-    return <>
-        <a onClick={toggleShowForm}>{title}</a>
-        {!!isVisible && <div className="micro-space">
-            <input ref={currentInput} value={asset.code || ''} onChange={onChangeCode} onKeyDown={onKeyDown} placeholder="Asset code"/>
-            <input value={asset.issuer || ''} onChange={onChangeIssuer} onKeyDown={onKeyDown} placeholder="Issuer address"/>
-            <div className="row micro-space">
-                <div className="column column-50">
-                    <Button block disabled={!isValid} onClick={onSave}>Save</Button>
-                </div>
-                <div className="column column-50">
-                    <Button block outline onClick={toggleShowForm}>Cancel</Button>
-                </div>
-            </div>
-        </div>}
-    </>
+    return <AddAssetEntryLayout title={title} currentInput={currentInput} asset={`${asset.code}:${asset.issuer}`}
+                                isEntered={isEntered} isValid={isValid} save={save}>
+        <input ref={currentInput} value={asset.code || ''} onChange={onChangeCode} onKeyDown={onKeyDown} placeholder="Asset code"/>
+        <input value={asset.issuer || ''} onChange={onChangeIssuer} onKeyDown={onKeyDown} placeholder="Issuer address"/>
+    </AddAssetEntryLayout>
 })
