+++ views/server-config/add-generic-asset-entry-form.js
@@ -2,16 +2,15 @@ import React, {useCallback, useRef, useState} from 'react'
 import {observer} from 'mobx-react'
 import AddAssetEntryLayout from './add-asset-entry-layout'
 
-export default observer(function AddSorobanTokenEntry({title, settings, save}) {
+export default observer(function AddGenericAssetEntry({title, settings, save}) {
     const [asset, setAsset] = useState('')
     const [isValid, setIsValid] = useState(false)
     const [isEntered, setIsEntered] = useState(false)
     const currentInput = useRef(null)
 
     const validate = useCallback(val => {
-        if (val.length !== 56 || !val.startsWith('C'))
-            return false
-        if (settings.data.assets.findIndex(asset => asset.code === val.toUpperCase()) !== -1) {
+        const pattern = new RegExp("[^a-zA-z]")
+        if (pattern.test(val) || settings.data.assets.findIndex(asset => asset.code === val.toUpperCase()) !== -1) {
             return false
         }
         return true
